requirements.txt
opencv-python
imutils
numpy
scipy
easyocr
pillow
mysql-connector-python
